reorganizando código entre projetos para estar de acordo com arquitetura hexagonal
