+++ src/core/SnackTech.Application/ModuleInjectionDependency.cs
@@ -1,7 +1,7 @@
 using System.Diagnostics.CodeAnalysis;
 using Microsoft.Extensions.DependencyInjection;
-using SnackTech.Application.Interfaces;
 using SnackTech.Application.UseCases;
+using SnackTech.Domain.Ports.Driven;
 
 namespace SnackTech.Application
 {
