+++ src/core/SnackTech.Application/UseCases/ProdutoService.cs
@@ -1,10 +1,10 @@
 using Microsoft.Extensions.Logging;
-using SnackTech.Application.Common;
-using SnackTech.Application.DTOs.Produto;
-using SnackTech.Application.Interfaces;
-using SnackTech.Domain.Contracts;
+using SnackTech.Domain.Common;
+using SnackTech.Domain.DTOs.Produto;
 using SnackTech.Domain.Guards;
 using SnackTech.Domain.Models;
+using SnackTech.Domain.Ports.Driven;
+using SnackTech.Domain.Ports.Driving;
 
 namespace SnackTech.Application.UseCases
 {
