+++ src/core/SnackTech.Domain/DTOs/Cliente/CadastroCliente.cs
@@ -1,5 +1,5 @@
 
-namespace SnackTech.Application.DTOs.Cliente
+namespace SnackTech.Domain.DTOs.Cliente
 {
     public class CadastroCliente
     {
