+++ src/core/SnackTech.Domain/DTOs/Pedido/RetornoPedido.cs
@@ -1,6 +1,6 @@
 using DomainModels = SnackTech.Domain.Models;
 
-namespace SnackTech.Application.DTOs.Pedido
+namespace SnackTech.Domain.DTOs.Pedido
 {
     public class RetornoPedido
     {
