+++ src/core/SnackTech.Domain/Ports/Driving/IClienteRepository.cs
@@ -1,6 +1,6 @@
 using SnackTech.Domain.Models;
 
-namespace SnackTech.Domain.Contracts
+namespace SnackTech.Domain.Ports.Driving
 {
     public interface IClienteRepository
     {
