+++ src/driven/SnackTech.Adapter.DataBase/ModuleInjectionDependency.cs
@@ -1,7 +1,7 @@
 using System.Diagnostics.CodeAnalysis;
 using Microsoft.Extensions.DependencyInjection;
 using SnackTech.Adapter.DataBase.Repositories;
-using SnackTech.Domain.Contracts;
+using SnackTech.Domain.Ports.Driving;
 
 namespace SnackTech.Adapter.DataBase
 {
